fix(ui): preserve silk aurora and add aurora flow
